requirements.txt
aiohttp>=3.9.1
aiodns>=3.1.1
aiosignal>=1.3.1
async-timeout>=4.0.3
attrs>=23.2.0

certifi>=2023.11.17
chardet>=5.2.0
click>=8.1.6
colorama>=0.4.6
cryptography>=41.0.7
idna>=3.6

Jinja2>=3.1.2
jsonschema>=4.10.3
packaging>=24.0
psutil>=5.9.8

Pygments>=2.17.2
PyYAML>=6.0.1
python-dateutil>=2.8.2
python-dotenv>=1.0.1
pytz>=2024.1
rich>=13.7.1
urllib3>=2.0.7
websocket-client>=1.7.0
yarl>=1.8.2

numpy>=1.26.4
sentence-transformers>=2.7.0
PySide6>=6.7.0
pillow>=10.3.0
pydantic>=2.7.0
mypy>=1.11.0
types-requests>=2.32.0.20240907
types-Pillow>=10.2.0.20240520
PySide6-stubs>=6.7.0

pytest>=8.3.0
pytest-cov>=5.0.0
ruff>=0.6.9
requests>=2.31.0
